Revert "Split builders and extends abstract & use trait"
This reverts commit 6ff9378d645598f36e38a7754a6b3b14a1319f1a.
